restructure + small adds
