+++ ascat/__init__.py
@@ -5,5 +5,5 @@ try:
 except:
     __version__ = 'unknown'
 
-from ascat.timeseries import *
+from ascat.read_native.cdr import *
 from ascat.h_saf import *
