+++ ascat/level2.py
@@ -35,9 +35,9 @@ from pygeobase.io_base import ImageBase
 from pygeobase.io_base import MultiTemporalImageBase
 from pygeobase.io_base import IntervalReadingMixin
 from pygeobase.object_base import Image
-import ascat.data_readers.read_eps as read_eps
-import ascat.data_readers.read_bufr as read_bufr
-import ascat.data_readers.read_nc as read_nc
+import ascat.read_native.eps_native as read_eps
+import ascat.read_native.bufr as read_bufr
+import ascat.read_native.nc as read_nc
 
 class AscatL2Image(ImageBase):
     """
